fix: Log annotation acquisition details file type data cannot be serialized
